Create a log request and response system.
Create a new class RequestAndResponseFilter that log requests and
responses data. To help this purpose the model of data logged is store
in classes RequestLogData and ResponseLogData.

The logic of printing the message is on classes that implements the new
interfaces RequestLogDataPrinter and ResponseLogDataPrinter, both are
functional interfaces so is possible use lambda expression instead of
implement this.

issues: ROL-11